requirements.txt
PyYAML
urllib
requests